requirements.txt
# Requirements for Indian News Verifier
# Core dependencies for the beautiful presentation app

streamlit>=1.28.0
google-generativeai>=0.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
